+++ tomato/symbolic/SymbTrAnalyzer.py
@@ -9,15 +9,17 @@ from symbtrextras.ScoreExtras import ScoreExtras
 from musicbrainzngs import NetworkError
 from musicbrainzngs import ResponseError
 
-from tomato.MCRCaller import MCRCaller
-from tomato.IO import IO
-from tomato.Analyzer import Analyzer
+from ..MCRCaller import MCRCaller
+from ..IO import IO
+from ..Analyzer import Analyzer
 
 # instantiate a mcr_caller
 _mcr_caller = MCRCaller()
 
 
 class SymbTrAnalyzer(Analyzer):
+    _inputs = ['boundaries', 'mbid', 'score_features']
+
     def __init__(self, verbose=False):
         super(SymbTrAnalyzer, self).__init__(verbose=verbose)
 
@@ -26,54 +28,69 @@ class SymbTrAnalyzer(Analyzer):
         self._mu2Reader = Mu2Reader()
         self._phraseSegmenter = _mcr_caller.get_binary_path('phraseSeg')
 
-    def analyze(self, txt_filepath, mu2_filepath, symbtr_name=None):
+    def analyze(self, txt_filepath, mu2_filepath, symbtr_name=None, **kwargs):
+        input_f = self._parse_inputs(**kwargs)
+
         # attempt to get the symbtr_name from the filename, if it is not given
         if symbtr_name is None:
             symbtr_name = os.path.splitext(os.path.basename(txt_filepath))[0]
 
         # Automatic phrase segmentation on the SymbTr-txt score
-        try:
-            # note this is already in 1-indexing which is the I/O convention
-            # of symbtrdataextractor>=v2.0.0-alpha.5
-            boundary_note_idx = self.segment_phrase(
-                txt_filepath, symbtr_name=symbtr_name)['boundary_note_idx']
-        except RuntimeError as e:
-            boundary_note_idx = None
-            warnings.warn(e.message, RuntimeWarning)
+        input_f['boundaries'] = self._call_analysis_step(
+            'segment_phrase', input_f['boundaries'], txt_filepath,
+            symbtr_name=symbtr_name)
 
-        # relevant recording or work mbid
-        mbid = self._get_first_mbid_from_symbtr_name(symbtr_name)
+        # get relevant recording or work mbid
+        # Note: very rare but there can be more that one mbid returned.
+        #       We are going to use the first mbid to fetch the metadata
+        # TODO: use all mbids
+        input_f['mbid'] = self._call_analysis_step(
+            'get_mbids', input_f['mbid'], symbtr_name)
+        if isinstance(input_f['mbid'], list):
+            input_f['mbid'] = input_f['mbid'][0]
 
-        # Extract the (meta)data from the SymbTr scores
+        # Extract the (meta)data from the SymbTr scores. Here the results from
+        # the previous steps are also summarized.
+        self._call_symbtr_data_extractor(txt_filepath, mu2_filepath,
+                                         symbtr_name, input_f)
+
+        return (input_f['score_features'], input_f['boundaries'],
+                input_f['mbid'])
+
+    def _call_symbtr_data_extractor(self, txt_filepath, mu2_filepath,
+                                    symbtr_name, features):
         try:
-            score_features, is_data_valid = self.extract_data(
-                txt_filepath, mu2_filepath, symbtr_name=symbtr_name, mbid=mbid,
-                segment_note_bound_idx=boundary_note_idx)
+            score_data = self._call_analysis_step(
+                'extract_data', features['score_features'],
+                txt_filepath, mu2_filepath, symbtr_name=symbtr_name,
+                mbid=features['mbid'], segment_note_bound_idx=features[
+                    'boundaries']['boundary_note_idx'])
         except (NetworkError, ResponseError):  # MusicBrainz is not available
             warnings.warn('Unable to reach http://musicbrainz.org/. '
                           'The metadata stored there is not crawled.',
                           RuntimeWarning)
-            score_features, is_data_valid = self.extract_data(
+            score_data = self._call_analysis_step(
+                'extract_data', features['score_features'],
                 txt_filepath, mu2_filepath, symbtr_name=symbtr_name,
-                segment_note_bound_idx=boundary_note_idx)
-
-        if not is_data_valid:
-            warnings.warn(symbtr_name + ' has validation problems.')
+                segment_note_bound_idx=features['boundaries'][
+                    'boundary_note_idx'])
+        if score_data is not None:
+            score_features, is_valid = score_data
+            if not is_valid:
+                warnings.warn(symbtr_name + ' has validation problems.')
+        else:
+            score_features, is_valid = [None, None]
 
         return score_features
 
     @staticmethod
-    def _get_first_mbid_from_symbtr_name(symbtr_name):
-        # Note: very rare but there can be more that one mbid returned.
-        #       We are going to use the first mbid to fetch the metadata
-        mbid = ScoreExtras.get_mbids(symbtr_name)
-        if not mbid:
+    def get_mbids(symbtr_name):
+        mbids = ScoreExtras.get_mbids(symbtr_name)
+        if not mbids:
             warnings.warn(u"No MBID returned for {0:s}".format(symbtr_name),
                           RuntimeWarning)
-        else:
-            mbid = mbid[0]
 
-        return mbid
+        return mbids
 
     def segment_phrase(self, txt_filename, symbtr_name=None):
         tic = timeit.default_timer()